pyproject.toml
[tool.poetry]
name = "aiolancium"
version = "0.2.0"
description = "AsyncIO Client for Lancium"
authors = ["Manuel Giffels <giffels@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers=[
        "Development Status :: 3 - Alpha",
        "Intended Audience :: Developers",
        "Intended Audience :: System Administrators",
        "Intended Audience :: Information Technology",
        "Intended Audience :: Science/Research",
        "Operating System :: OS Independent",
        "Topic :: Internet :: WWW/HTTP :: Session",
        "Topic :: Utilities",
        "Framework :: AsyncIO",
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3.6",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
    ]
keywords=["asyncio", "lancium", "compute client"]
packages = [
    { include = "aiolancium" }
]

[tool.poetry.dependencies]
python = "^3.6.2"
simple-rest-client = "0.5.4"  # version to support aiohttp, matching AsyncOpenStack client. client cannot be re-used with httpx.
jsonref = "^1.0.0"
flake8-bugbear = {version = "22.9.23", optional = true}
black = {version = "<=22.8.0", optional = true}
aioresponses = {version = "^0.7.3", optional = true}
flake8 = {version = "^5.0.4", optional = true}
sphinx-rtd-theme = {version = "^1.1.1", optional = true}
sphinxcontrib-contentui = {version = "^0.2.5", optional = true}
sphinx = {version = "4.3.1", optional = true}
pyjwt = "<=2.4.0"

[tool.poetry.extras]
doc = [
    "sphinx",
    "sphinx_rtd_theme",
    "sphinxcontrib-contentui",
]
test = [
    "flake8",
    "flake8-bugbear",
    "black",
    "aioresponses"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
